requirements.txt
djangorestframework
asgiref==3.9.1
sqlparse==0.5.3
djangorestframework
djangorestframework-simplejwt
drf-yasg==1.21.10
django_environ==0.11.2
psycopg2-binary==2.9.10
whitenoise==6.8.2
pillow==11.0.0
gunicorn==21.2.0
celery==5.3.6
django-celery-beat==2.5.0
redis==5.0.1
django-jazzmin==3.0.1
django-cors-headers
django-tailwind==3.8.0
django-browser-reload==1.12.1